Ensure that window-level toolbars and menus are created on secondary MainWindows.
